+++ include/heyoka/detail/dfloat.hpp
@@ -40,6 +40,7 @@ struct dfloat {
     explicit dfloat(F x) : hi(x), lo(0) {}
     // NOLINTNEXTLINE(bugprone-easily-swappable-parameters)
     explicit dfloat(F h, F l) : hi(h), lo(l) {}
+    explicit dfloat(std::pair<F, F> p) : dfloat(std::move(p.first), std::move(p.second)) {}
 
     explicit operator F() const
     {