+++ include/heyoka/detail/erfa_decls.hpp
@@ -18,6 +18,8 @@ int eraUtcut1(double, double, double, double *, double *);
 int eraTaiutc(double, double, double *, double *);
 double eraEra00(double, double);
 int eraCal2jd(int, int, int, double *, double *);
+int eraJd2cal(double, double, int *, int *, int *, double *);
+int eraTttai(double, double, double *, double *);
 }
 
 #endif