+++ tools/gha_osx_x86.sh
@@ -13,7 +13,7 @@ export PATH="$HOME/miniconda/bin:$PATH"
 bash miniconda.sh -b -p $HOME/miniconda
 conda create -y -p $deps_dir c-compiler zlib cxx-compiler libcxx cmake ninja \
     llvmdev tbb-devel tbb libboost-devel sleef xtensor xtensor-blas blas openssl \
-    blas-devel fmt spdlog mppp 'clang<19' 'clangxx<19'
+    blas-devel fmt spdlog mppp clang clangxx
 source activate $deps_dir
 
 # Create the build dir and cd into it.
@@ -25,6 +25,10 @@ unset CXXFLAGS
 unset CFLAGS
 
 # Configure.
+#
+# NOTE: regarding D_LIBCPP_DISABLE_AVAILABILITY, see:
+#
+# https://conda-forge.org/docs/maintainer/knowledge_base/#newer-c-features-with-old-sdk
 CXX=clang++ CC=clang cmake -G Ninja ../ \
     -DCMAKE_PREFIX_PATH=$deps_dir \
     -DCMAKE_BUILD_TYPE=Debug \
@@ -32,13 +36,16 @@ CXX=clang++ CC=clang cmake -G Ninja ../ \
     -DHEYOKA_WITH_MPPP=yes \
     -DHEYOKA_BUILD_TUTORIALS=ON \
     -DHEYOKA_WITH_SLEEF=yes \
-    -DCMAKE_CXX_FLAGS_DEBUG="-g -Og"
+    -DCMAKE_CXX_FLAGS_DEBUG="-g -Og" \
+    -DCMAKE_CXX_FLAGS="-D_LIBCPP_DISABLE_AVAILABILITY -fsanitize=address"
 
 # Build.
 ninja -v
 
 # Run the tests.
-ctest -VV -j4
+#
+# See: https://github.com/google/sanitizers/wiki/AddressSanitizerContainerOverflow
+ASAN_OPTIONS=detect_container_overflow=0 ctest -VV -j4
 
 set +e
 set +x